refactor: Moved position function to own file
